# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Cardone Loans & Grants API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /applications:
    get:
      operationId: getApplications
      tags: [applications]
      summary: Get user's applications
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createApplication
      tags: [applications]
      summary: Submit a new application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApplicationRequest"
      responses:
        "201":
          description: Application created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/applications:
    get:
      operationId: adminGetApplications
      tags: [admin]
      summary: Admin - get all applications
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [loan, grant]
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, under_review, approved, rejected]
      responses:
        "200":
          description: All applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApplicationWithUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/applications/{id}/approve:
    post:
      operationId: approveApplication
      tags: [admin]
      summary: Admin - approve application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminActionRequest"
      responses:
        "200":
          description: Application approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/applications/{id}/reject:
    post:
      operationId: rejectApplication
      tags: [admin]
      summary: Admin - reject application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminActionRequest"
      responses:
        "200":
          description: Application rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/applications/{id}/update:
    patch:
      operationId: adminUpdateApplication
      tags: [admin]
      summary: Admin - update application details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateApplicationRequest"
      responses:
        "200":
          description: Application updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/payments:
    get:
      operationId: adminGetPayments
      tags: [admin]
      summary: Admin - get all manual payments
      responses:
        "200":
          description: All manual payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ManualPayment"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/payments/{id}/verify:
    post:
      operationId: adminVerifyPayment
      tags: [admin]
      summary: Admin - verify a manual payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManualPayment"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/stats:
    get:
      operationId: adminGetStats
      tags: [admin]
      summary: Admin - get dashboard statistics
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    CreateApplicationRequest:
      type: object
      properties:
        type:
          type: string
          enum: [loan, grant]
        category:
          type: string
          enum: [personal, business]
        amountRequested:
          type: number
        fullName:
          type: string
        nationalId:
          type: string
        phoneNumber:
          type: string
        employmentStatus:
          type: string
        monthlyIncome:
          type: number
        purposeOfFunds:
          type: string
        businessName:
          type: string
        registrationNumber:
          type: string
        kraPin:
          type: string
        businessType:
          type: string
        annualRevenue:
          type: number
        ownerDetails:
          type: string
      required:
        - type
        - category
        - amountRequested
        - purposeOfFunds
    Application:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [loan, grant]
        category:
          type: string
          enum: [personal, business]
        amountRequested:
          type: number
        preapprovedAmount:
          type: number
          nullable: true
        status:
          type: string
          enum: [pending, under_review, approved, rejected]
        paymentCode:
          type: string
          nullable: true
        fullName:
          type: string
          nullable: true
        nationalId:
          type: string
          nullable: true
        phoneNumber:
          type: string
          nullable: true
        employmentStatus:
          type: string
          nullable: true
        monthlyIncome:
          type: number
          nullable: true
        purposeOfFunds:
          type: string
          nullable: true
        businessName:
          type: string
          nullable: true
        registrationNumber:
          type: string
          nullable: true
        kraPin:
          type: string
          nullable: true
        businessType:
          type: string
          nullable: true
        annualRevenue:
          type: number
          nullable: true
        ownerDetails:
          type: string
          nullable: true
        adminComment:
          type: string
          nullable: true
        reason:
          type: string
          nullable: true
        approvedAmount:
          type: number
          nullable: true
        assignedPartner:
          type: string
          nullable: true
        disbursementDate:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - type
        - category
        - amountRequested
        - status
        - createdAt
        - updatedAt
    ApplicationWithUser:
      allOf:
        - $ref: "#/components/schemas/Application"
        - type: object
          properties:
            userEmail:
              type: string
            userFullName:
              type: string
          required:
            - userEmail
            - userFullName
...
    AdminActionRequest:
      type: object
      properties:
        reason:
          type: string
      required:
        - reason
    AdminUpdateApplicationRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, under_review, approved, rejected]
        approvedAmount:
          type: number
          nullable: true
        assignedPartner:
          type: string
          nullable: true
        disbursementDate:
          type: string
          nullable: true
        adminComment:
          type: string
          nullable: true
    ManualPayment:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        applicationId:
          type: integer
          nullable: true
        userName:
          type: string
        phoneNumber:
          type: string
        mpesaConfirmationCode:
          type: string
        amountKes:
          type: number
        isVerified:
          type: boolean
        verifiedBy:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - userName
        - phoneNumber
        - mpesaConfirmationCode
        - amountKes
        - isVerified
        - createdAt
    AdminStats:
      type: object
      properties:
        totalApplications:
          type: integer
        pendingApplications:
          type: integer
        approvedApplications:
          type: integer
        rejectedApplications:
          type: integer
        underReviewApplications:
          type: integer
        totalLoans:
          type: integer
        totalGrants:
          type: integer
        recentApplications:
          type: array
          items:
            $ref: "#/components/schemas/ApplicationWithUser"
      required:
        - totalApplications
        - pendingApplications
        - approvedApplications
        - rejectedApplications
        - underReviewApplications
        - totalLoans
        - totalGrants
        - recentApplications