# lib/api-zod/src/generated/types/application.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cardone Loans & Grants API
 * OpenAPI spec version: 0.1.0
 */
import type { ApplicationCategory } from "./applicationCategory";
import type { ApplicationStatus } from "./applicationStatus";
import type { ApplicationType } from "./applicationType";

export interface Application {
  id: number;
  userId: number;
  type: ApplicationType;
  category: ApplicationCategory;
  amountRequested: number;
  preapprovedAmount?: number | null;
  status: ApplicationStatus;
  paymentCode?: string | null;
  fullName?: string | null;
  nationalId?: string | null;
  phoneNumber?: string | null;
  employmentStatus?: string | null;
  monthlyIncome?: number | null;
  purposeOfFunds?: string | null;
  businessName?: string | null;
  registrationNumber?: string | null;
  kraPin?: string | null;
  businessType?: string | null;
  annualRevenue?: number | null;
  ownerDetails?: string | null;
  adminComment?: string | null;
  reason?: string | null;
  approvedAmount?: number | null;
  assignedPartner?: string | null;
  disbursementDate?: string | null;
  createdAt: Date;
  updatedAt: Date;
}